Fix bug.

+++ src/main/java/pl/edu/wat/lab/usercontacts/exception/contact/ContactNotFoundException.java
@@ -2,7 +2,7 @@ package pl.edu.wat.lab.usercontacts.exception.contact;
 
 public class ContactNotFoundException extends RuntimeException {
 
-    public ContactNotFoundException(int contactId) {
+    public ContactNotFoundException(Long contactId) {
         super("Could not find contact with id = " + contactId);
     }
 }
